refactor: AuthenticationInterceptor에 템플릿메서드패턴 구현
