+++ src/main/java/nextstep/subway/auth/ui/session/SessionAuthenticationInterceptor.java
@@ -7,40 +7,26 @@ import nextstep.subway.auth.domain.UserDetails;
 import nextstep.subway.auth.exception.InvalidPasswordException;
 import nextstep.subway.auth.exception.NoUserFoundException;
 import nextstep.subway.auth.infrastructure.SecurityContext;
+import nextstep.subway.auth.ui.AbstractAuthenticationInterceptor;
 import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
+import java.io.IOException;
 import java.util.Map;
 
 import static nextstep.subway.auth.infrastructure.SecurityContextHolder.SPRING_SECURITY_CONTEXT_KEY;
 
-public class SessionAuthenticationInterceptor implements HandlerInterceptor {
+public class SessionAuthenticationInterceptor extends AbstractAuthenticationInterceptor {
     public static final String USERNAME_FIELD = "username";
     public static final String PASSWORD_FIELD = "password";
 
-    private CustomUserDetailsService userDetailsService;
-
-    public SessionAuthenticationInterceptor(CustomUserDetailsService userDetailsService) {
-        this.userDetailsService = userDetailsService;
+    public SessionAuthenticationInterceptor(CustomUserDetailsService customUserDetailsService) {
+        super(customUserDetailsService);
     }
 
     @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
-        AuthenticationToken token = convert(request);
-        Authentication authentication = authenticate(token);
-
-        if (authentication == null) {
-            throw new RuntimeException();
-        }
-
-        HttpSession httpSession = request.getSession();
-        httpSession.setAttribute(SPRING_SECURITY_CONTEXT_KEY, new SecurityContext(authentication));
-        response.setStatus(HttpServletResponse.SC_OK);
-        return false;
-    }
-
     public AuthenticationToken convert(HttpServletRequest request) {
         Map<String, String[]> paramMap = request.getParameterMap();
         String principal = paramMap.get(USERNAME_FIELD)[0];
@@ -49,21 +35,10 @@ public class SessionAuthenticationInterceptor implements HandlerInterceptor {
         return new AuthenticationToken(principal, credentials);
     }
 
-    public Authentication authenticate(AuthenticationToken token) {
-        String principal = token.getPrincipal();
-        UserDetails userDetails = userDetailsService.loadUserByUsername(principal);
-        checkAuthentication(userDetails, token);
-
-        return new Authentication(userDetails);
-    }
-
-    private void checkAuthentication(UserDetails userDetails, AuthenticationToken token) {
-        if (userDetails == null) {
-            throw new NoUserFoundException();
-        }
-
-        if (!userDetails.checkPassword(token.getCredentials())) {
-            throw new InvalidPasswordException();
-        }
+    @Override
+    protected void setResponse(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
+        HttpSession httpSession = request.getSession();
+        httpSession.setAttribute(SPRING_SECURITY_CONTEXT_KEY, new SecurityContext(authentication));
+        response.setStatus(HttpServletResponse.SC_OK);
     }
 }
