+++ src/main/java/nextstep/subway/auth/ui/token/TokenAuthenticationInterceptor.java
@@ -11,35 +11,23 @@ import nextstep.subway.auth.dto.TokenResponse;
 import nextstep.subway.auth.exception.InvalidPasswordException;
 import nextstep.subway.auth.exception.NoUserFoundException;
 import nextstep.subway.auth.infrastructure.JwtTokenProvider;
+import nextstep.subway.auth.ui.AbstractAuthenticationInterceptor;
 import org.springframework.http.MediaType;
-import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class TokenAuthenticationInterceptor implements HandlerInterceptor {
-
-    private final CustomUserDetailsService customUserDetailsService;
+public class TokenAuthenticationInterceptor extends AbstractAuthenticationInterceptor {
     private final JwtTokenProvider jwtTokenProvider;
     private final ObjectMapper objectMapper = new ObjectMapper();
 
     public TokenAuthenticationInterceptor(CustomUserDetailsService customUserDetailsService, JwtTokenProvider jwtTokenProvider) {
-        this.customUserDetailsService = customUserDetailsService;
+        super(customUserDetailsService);
         this.jwtTokenProvider = jwtTokenProvider;
     }
 
     @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws IOException {
-        AuthenticationToken authenticationToken = convert(request);
-        Authentication authentication = authenticate(authenticationToken);
-        TokenResponse tokenResponse = convertAuthToToken(authentication);
-
-        setResponse(response, tokenResponse);
-
-        return false;
-    }
-
     public AuthenticationToken convert(HttpServletRequest request) throws IOException {
         TokenRequest tokenRequest = objectMapper.readValue(request.getInputStream(), TokenRequest.class);
         String principal = tokenRequest.getEmail();
@@ -48,33 +36,18 @@ public class TokenAuthenticationInterceptor implements HandlerInterceptor {
         return new AuthenticationToken(principal, credentials);
     }
 
-    public Authentication authenticate(AuthenticationToken authenticationToken) {
-        String principal = authenticationToken.getPrincipal();
-        UserDetails userDetails = customUserDetailsService.loadUserByUsername(principal);
-        checkAuthentication(userDetails, authenticationToken);
-        return new Authentication(userDetails);
-    }
-
-    private void checkAuthentication(UserDetails userDetails, AuthenticationToken authenticationToken) {
-        if(userDetails == null) {
-            throw new NoUserFoundException();
-        }
-
-        if (!userDetails.checkPassword(authenticationToken.getCredentials())) {
-            throw new InvalidPasswordException();
-        }
+    @Override
+    protected void setResponse(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
+        TokenResponse tokenResponse = convertAuthToToken(authentication);
+        String responseToClient = new ObjectMapper().writeValueAsString(tokenResponse);
+        response.setStatus(HttpServletResponse.SC_OK);
+        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+        response.getOutputStream().print(responseToClient);
     }
 
     private TokenResponse convertAuthToToken(Authentication authentication) throws JsonProcessingException {
         String payload = objectMapper.writeValueAsString(authentication.getPrincipal());
         String token = jwtTokenProvider.createToken(payload);
         return new TokenResponse(token);
-    }
-
-    private void setResponse(HttpServletResponse response, TokenResponse tokenResponse) throws IOException {
-        String responseToClient = new ObjectMapper().writeValueAsString(tokenResponse);
-        response.setStatus(HttpServletResponse.SC_OK);
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.getOutputStream().print(responseToClient);
     }
 }
